# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Nutritionist API
  version: 0.1.0
  description: |
    HTTP interface for the AI Nutritionist platform. Endpoints power the React Native app,
    web console, and conversational channels.
servers:
  - url: https://api.ai-nutritionist.test/v1
paths:
  /auth/login:
    post:
      summary: Begin authentication via magic link / OTP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, identifier]
              properties:
                channel:
                  type: string
                  enum: [sms, email]
                identifier:
                  type: string
                  description: Phone number (E.164) or email address.
      responses:
        '202':
          description: Challenge delivered.
...
  /me:
    get:
      summary: Retrieve current user profile and entitlements.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserProfile:
      type: object
      required: [id]
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        dietaryPreferences:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
        budgetPerWeek:
          type: number
          nullable: true
        goals:
          type: array
          items:
            $ref: '#/components/schemas/GoalPreference'
    GoalPreference:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [weight_loss, weight_gain, maintenance, performance]
        targetCalories:
          type: number
        proteinTarget:
          type: number
          nullable: true
...